[#104] refactor : Member 서비스 계층 helper로 책임 분리 + 주석 작성
